API

Model

AgentBasedModels.Model — Type

mutable struct Model

Basic structure which contains the agent based model.

Examples

m = Model(); #Create empty model

addGlobal!(m,:x); #Add a global variable to the model
# Here there may be many more additions

Add Parameters

AgentBasedModels.addGlobal! — Function

function addGlobal!(agentModel::Model, addvar::Symbol; updates="", randVar = Tuple{Symbol,String}[])

Add a global variable to the model with optional update rules.

Examples

m = Model();

addGlobal!(m,:x);

m = Model();
update = "
x= r #r is a random variable with μ=0. and σ=1.
"

addGlobal!(m,:x,updates=update,randVar=[(:r,Normal(0.,1.))]);

function addGlobal!(agentModel::Model, addvar::Array{Symbol}; updates="", randVar = Tuple{Symbol,String}[])

Add a set of global variables to the model with optional update rules.

Examples

m = Model();

addGlobal!(m,[:x,:y]);

m = Model();
update = " #We update x but not y
x= r
"

addGlobal!(m,[:x,:y],updates=update,randVar=[(:r,Normal,0,1)]);

AgentBasedModels.addLocal! — Function

addLocal!(agentModel::Model, addvar::Symbol; updates="", randVar = Tuple{Symbol,String}[])

Add a local variable to the model with optional update rules.

Examples

m = Model();

addLocal!(m,:x);

m = Model();
update = "
x= r #r is a random variable with μ=0. and σ=1.
"

addLocal!(m,:x,updates=update,randVar=[(:r,Normal,0.,1.)]);

addLocal!(agentModel::Model, addvar::Array{Symbol}; updates="", randVar = Tuple{Symbol,String}[])

Add a local variable to the model with optional update rules.

Examples

m = Model();

addGlobal!(m,[:x,:y]);

m = Model();
update = " #We update x but not y
x= r
"

addGlobal!(m,[:x,:y],updates=update,randVar=[(:r,Normal,0,1)]);

AgentBasedModels.addLocalInteraction! — Function

function addLocalInteraction!(agentModel::Model, addvar::Symbol, addeqs::String; randVar = Tuple{Symbol,String}[])

Add a local interaction to the model.

Examples

m = Model();
addLocal!(m,[:x,:y]);

interaction = "
d₁ = sqrt((x₁-x₂)^2+(y₁-y₂)^2)
"
addLocalInteraction!(m,:d,interaction);

function addLocalInteraction!(agentModel::Model, addvar::Symbol, addeqs::String; randVar = Tuple{Symbol,String}[])

Add a local interaction to the model.

Examples

m = Model();
addLocal!(m,[:x,:y]);

interaction = "
d₁ = sqrt((x₁-x₂)^2+(y₁-y₂)^2)
dAbs₁ = abs(x₁-x₂)+abs(y₁-y₂)
"
addLocalInteraction!(m,[:d,:dAbs],interaction);

AgentBasedModels.addVariable! — Function

function addVariable!(agentModel::Model, addvar::Symbol, addeqs::String)

Add a variable described by an Ordinary or Stochastic Differential Equation.

Examples

m = Model();
eq = "
dxdt = -x #Exponential decay
"
addVariable!(m,:x,eq);

m = Model();
eq = "
dxdt = -x+ξ #Wiener process
"
addVariable!(m,:x,eq);

function addVariable!(agentModel::Model, addvar::Symbol, addeqs::String)

Add a variable described by an Ordinary or Stochastic Differential Equations.

Examples

m = Model();
eq = "
dxdt = -x #Exponential decay
dydt = -y + ξ #Wiener process
"
addVariable!(m,[:x,:y],eq);

AgentBasedModels.addInteraction! — Function

function addInteraction!(agentModel::Model, addvar::Symbol, addeqs::String)

Add a interaction parameters to the model. Differently to the local interactions, this parameters will be updated inside the integration steps of the differential equations.

Examples

m = Model();
eq = "
dxdt = -x+g*ξ #Wiener process with an additional interaction
"
addVariable!(m,:x,eq);

interaction = "
g₁ += 1./sqrt((x₁-x₂)^2+(y₁-y₂)^2) #The difussion will be affected by the presence of other particles around them.
"
addInteraction!(m,:g,interaction);

function addInteraction!(agentModel::Model, addvar::Symbol, addeqs::String)

Add a interaction parameters to the model. Differently to the local interactions, this parameters will be updated inside the integration steps of the differential equations.

Examples

m = Model();
eq = "
dxdt = -x+g*ξ+p #Wiener process with an additional interaction
"
addVariable!(m,:x,eq);

interaction = "
g₁ += 1./sqrt((x₁-x₂)^2+(y₁-y₂)^2) #The difussion will be affected by the presence of other particles around them.
p₁ += 1./(abs(x₁-x₂)+abs(y₁-y₂)) #The difussion will be affected by the presence of other particles around them.
"
addInteraction!(m,[:g,:p],interaction);

Neighborhoods

Missing docstring.

Missing docstring for setNeighborhoodFull!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for setNeighborhoodAdjacency!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for setNeighborhoodGrid!. Check Documenter's build log for details.

Special

AgentBasedModels.addDivision! — Function

function addDivision!(agentModel::Model, condition::String, update::String; randVar = Tuple{Symbol,String}[])

Function that adds a division process of the particles to the model. Cells divide under condition and update the new parameters with update.

Examples

m = Model();
addLocal!(m,[:g,:tDiv]); #Add two local variables, g and the time of division tDiv

condition = 
"
tDiv < t
"

update =
"
g₁ = gₚ*r #Asymmetric split of the component between the two daughter particles
g₁ = tDivₚ+1.

g₂ = gₚ*(1-r)
g₂ = tDivₚ+1.
"

addDivision!(m,condition,update,randVar=[(:r,"Uniform",0.,1.)])

AgentBasedModels.addPseudopode! — Function

function addPseudopode!(agentModel::Model, var::Symbol, tVar::Symbol, neighbourCondition::String, force::String, updateChange::String; randVar = Tuple{Symbol,String}[])

Add a pseudopode force.

Examples

m = Model()
addLocal!([:x,:y])

neighbourCondition = 
"
sqrt((x₁-x₂)^2+(y₁-y₂)^2) < 2.
"
force = 
"
f = sqrt((x₁-x₂)^2+(y₁-y₂)^2)*exp(-sqrt((x₁-x₂)^2+(y₁-y₂)^2))
"
updateChange = 
"
tPseudo = t + σPseudo
"
addPseudopode!(m, :f, :tPseudo, condition, force, updateChange, randVar = [(:σPseudo,"Uniform",1.,2.)])