API for Developers

Auxiliar functions

This is a compendium of handful functions that are used all the time during the adaptation of the rest of the models.

AgentBasedModels.platformAdapt — Function

Function to adapt block of code into the CPU of GPU platforms.

Arguments

text (String, Expr, Array{String}, Array{Expr}) Block of code to adapt.

Optional keyword arguments

platform (String) Platform to be adapted. Options are "cpu" (default) or "gpu".

Returns

Expr or Array{Expr}

AgentBasedModels.vectParams — Function

Function to vectorize the variables in a block of code.

Arguments

agentModel (Model) Model that is being compiled.

text (String, Expr, Array{String}, Array{Expr}) Block(s) of code to be vectorized according to the parameters stored in the agentModel.

Returns

Expr or Array{Expr}

AgentBasedModels.adapt — Function

Function that performs the vectParams, neighbourhAdapt and platformAdapt in succession.

Arguments

agentModel (Model) Model that is being compiled.

text (String, Expr) Expression or expressions to be vectorized according to the parameters stored in the agentModel.

platform (String) Platform to be adapted. Options are "cpu" (default) or "gpu".

Returns

Expr

AgentBasedModels.pushAdapt! — Function

Function that pushes in a container a text after adaptation.

Arguments

container (Array) Array where the adapted object is stored

agentModel (Model) Model that is being compiled.

text (String, Expr) Expression or expressions to be vectorized according to the parameters stored in the agentModel.

platform (String) Platform to be adapted. Options are "cpu" (default) or "gpu".

Returns

nothing

AgentBasedModels.addIfNot! — Function

Add an element if it is not present in an Array

Arguments

container (Array) Container where to introduce the elements

object (Any or Array{Any}) Array of elements

Returns

nothing

AgentBasedModels.checkDeclared — Function

Function that checks if an Symbol has already been declared in the model and returns an error if it is duplicated.

Arguments

agentModel (Model) Agent Model

s (Symbol, Array{Symbol}) Symbol(s) to be checked for duplicated declaration.

Returns

nothing

AgentBasedModels.clean — Function

Cleans a complex expression. It is very helpful for debuging the compiled model evolve function after parsing everything.

Arguments

a (Expr) Expression to be cleaned

Returns

nothing

AgentBasedModels.commonArguments — Function

Returns list of all vectorized parameters of the model added by basic functions.

Arguments

agentModel (Model) Agent Model

Optional keywork Arguments

random (Bool) If true (default), returns also symbols for the vectorized random variables.

Return

Array{Symbol}

AgentBasedModels.findSymbol — Function

Check if a symbol or expression is present in a block of code

Arguments

exp (Expr, String) Block of code where to check for the symbol

ob (Symbol or Expr) Symbol or Expression to look for in the block of code

Returns

Bool

AgentBasedModels.splits — Function

Recurrent function auxiliar to the splitEqs for the block of code adaptation

AgentBasedModels.splitEqs — Function

Prepares a block of code with differential equations to be vectorized and introduced in the integration algorithms. Returns the adapted block of code and the list of the lines where there has been declared a Wienner random variable.

Arguments

value (Expr) Block of code with differential equations

Returns

Expr, Array{Int}

AgentBasedModels.subs — Function

Substitution of a symbol by another symbol in a block of code.

Arguments

exp (String, Expr, Array{String}, Array{Expr}) Block of code where to substitute the symbol

ob (String, Expr, Array{String}, Array{Expr}) Object to be substituted

tar (String, Expr, Array{String}, Array{Expr}) Target by which is substituted, it must have the same type than ob.

Returns

Expr or Array{Expr}

AgentBasedModels.parameterAdapt — Function

function parameterAdapt(agentModel::Model,inLoop,arg;platform::String="cpu",nChange_=false)

Function that returns the pieces of the final compiling code for the parameters adapted to the corresponding platform:

arrays to declare containg the parameters declared

functions for parameter updates

execution lines

Parameters:

agentModel : Model structure

inLoop : Code of the interaction local to be adapted depending on the neighbborhood

arg : Additional arguments required by the functions

platform : Platform to be adapted ("cpu" or "gpu")

nChange_ : FILL THE GAP

Precompilation

AgentBasedModels.precompile! — Function

Main function that plugs in all the declared parts of the Agent Based Model and generates an evolution function. It is exactly the same as the compile function but it evals in the local scope of the module. Used for compilationof the predefined models of the library.

Arguments

agentModel (Model) Agent Model to be compiled.

Optative keywork arguments

integrator (String) Integrator to be implemented in the model ("euler" by default)

saveRAM (Bool) Indicate if the steps have to be saved in a CommunityInTime structure. False by default.

saveVTK (Bool) Indicate if the steps have to be saved in a VTK file (experimental). False by default.

positionsVTK (Array{Symbols}) The declared symbols that will correspond to the VTK spatial positions. [:x,:y,:z] by default.

debug (Bool) Print the cleaned compiled function for debugging purposes. False by default.

Returns

nothing

Set neighbourhood

All the neighborhood types X share the same structure. They contain always four elements:

A NeighborhoodX structure that keeps all the required elements in order to define the neighborhood. This structured is created when calling the setNeighborhoodX! and added to the model.

The setNeighborhoodX! function that is the high level function for seting a neighborhood. The details are present in the API.

A neighboursX function that is runned during compilation time and creates the functions that computed to make the neighborhoods.

A neighboursXAdapt that assigns the inner loop iterator :nnic2_ into the appropiate term depending on the neighborhood type.

Functions defined in 2 and 4 have always the same inputs and outputs.

Full connected

AgentBasedModels.NeighboursFull — Type

struct NeighboursFull <: Neighbours

AgentBasedModels.neighboursFull — Function

function neighboursFull(agentModel::Model;platform="cpu")

AgentBasedModels.neighboursFullAdapt — Function

function neighboursFullAdapt(entry)

By Adjacecy

AgentBasedModels.NeighboursAdjacency — Type

struct NeighboursAdjacency <: Neighbours

AgentBasedModels.neighboursByAdjacency — Function

function neighboursByAdjacency(agentModel::Model;platform="cpu")

AgentBasedModels.neighboursByAdjacencyAdapt — Function

function neighboursByAdjacencyAdapt(entry)

By Grid

AgentBasedModels.NeighboursGrid — Type

struct NeighboursGrid <: Neighbours

AgentBasedModels.neighboursByGrid — Function

function neighboursByGrid(agentModel::Model;platform="cpu")

AgentBasedModels.neighboursByGridAdapt — Function

function neighboursByGridAdapt(entry)

AgentBasedModels.loopNeighbourGridCreation — Function

function loopNeighbourGridCreation(i,i0,n,x=nothing,pos="")

Auxiliar function for creating nested loops during the grid creation.

Integrators

AgentBasedModels.integratorEuler — Function

Adaptation to the Euler integrator in Îto prescription.

\[x(t+Δt) = x(t) + f(x(t),t)*Δt + g(x(t),t)ΔW\]

where $ΔW$ is a Wiener process with step proportional to $Δt^{1/2}$.

Arguments

agentModel (Model) Agent Model

inLoop (Expr) Block of code that will be included in the integration step.

arg (Array{Symbol} Additional arguments to include from the neighborhood.

Optatiove keywork arguments

platform (String) Platform to be adapted the algorithm. "cpu" by default

AgentBasedModels.integratorHeun — Function

Adaptation to the Stochastic Improved Euler integrator in Îto prescription.

\[K₁ = f(x(t),t)*Δt + g(x(t),t)(ΔW-S Δt^{1/2} )\]

\[K₂ = f(x(t)+K₁,t+Δt)*Δt + g(x(t)+K₁,t+Δt)(ΔW+SΔt^{1/2} )\]

\[x(t+Δt) = x(t) + (K₁+K₂)/2\]

where $S$ is a random variable chosen to be ±1 with probability 1/2 and $ΔW$ is a Wiener process with step proportional to $Δt^{1/2}$.

Arguments

agentModel (Model) Agent Model

inLoop (Expr) Block of code that will be included in the integration step.

arg (Array{Symbol} Additional arguments to include from the neighborhood.

Optatiove keywork arguments

platform (String) Platform to be adapted the algorithm. "cpu" by default