API reference

Model

AgentModel.Model — Type

mutable struct Model

Basic structure which contains the agent based model.

Examples

m = Model(); #Create empty model

addGlobal!(m,:x); #Add a global variable to the model
# Here there may be many more additions

Add Parameters

AgentModel.addGlobal! — Function

function addGlobal!(agentModel::Model, addvar::Symbol; updates="", randVar = Tuple{Symbol,String}[])

Add a global variable to the model with optional update rules.

Examples

m = Model();

addGlobal!(m,:x);

m = Model();
update = "
x= r #r is a random variable with μ=0. and σ=1.
"

addGlobal!(m,:x,updates=update,randVar=[(:r,Normal(0.,1.))]);

function addGlobal!(agentModel::Model, addvar::Array{Symbol}; updates="", randVar = Tuple{Symbol,String}[])

Add a set of global variables to the model with optional update rules.

Examples

m = Model();

addGlobal!(m,[:x,:y]);

m = Model();
update = " #We update x but not y
x= r
"

addGlobal!(m,[:x,:y],updates=update,randVar=[(:r,Normal,0,1)]);

AgentModel.addLocal! — Function

addLocal!(agentModel::Model, addvar::Symbol; updates="", randVar = Tuple{Symbol,String}[])

Add a local variable to the model with optional update rules.

Examples

m = Model();

addLocal!(m,:x);

m = Model();
update = "
x= r #r is a random variable with μ=0. and σ=1.
"

addLocal!(m,:x,updates=update,randVar=[(:r,Normal,0.,1.)]);

addLocal!(agentModel::Model, addvar::Array{Symbol}; updates="", randVar = Tuple{Symbol,String}[])

Add a local variable to the model with optional update rules.

Examples

m = Model();

addGlobal!(m,[:x,:y]);

m = Model();
update = " #We update x but not y
x= r
"

addGlobal!(m,[:x,:y],updates=update,randVar=[(:r,Normal,0,1)]);

AgentModel.addLocalInteraction! — Function

function addLocalInteraction!(agentModel::Model, addvar::Symbol, addeqs::String; randVar = Tuple{Symbol,String}[])

Add a local interaction to the model.

Examples

m = Model();
addLocal!(m,[:x,:y]);

interaction = "
d₁ = sqrt((x₁-x₂)^2+(y₁-y₂)^2)
"
addLocalInteraction!(m,:d,interaction);

function addLocalInteraction!(agentModel::Model, addvar::Symbol, addeqs::String; randVar = Tuple{Symbol,String}[])

Add a local interaction to the model.

Examples

m = Model();
addLocal!(m,[:x,:y]);

interaction = "
d₁ = sqrt((x₁-x₂)^2+(y₁-y₂)^2)
dAbs₁ = abs(x₁-x₂)+abs(y₁-y₂)
"
addLocalInteraction!(m,[:d,:dAbs],interaction);

AgentModel.addVariable! — Function

function addVariable!(agentModel::Model, addvar::Symbol, addeqs::String)

Add a variable described by an Ordinary or Stochastic Differential Equation.

Examples

m = Model();
eq = "
dxdt = -x #Exponential decay
"
addVariable!(m,:x,eq);

m = Model();
eq = "
dxdt = -x+ξ #Wiener process
"
addVariable!(m,:x,eq);

function addVariable!(agentModel::Model, addvar::Symbol, addeqs::String)

Add a variable described by an Ordinary or Stochastic Differential Equations.

Examples

m = Model();
eq = "
dxdt = -x #Exponential decay
dydt = -y + ξ #Wiener process
"
addVariable!(m,[:x,:y],eq);

AgentModel.addInteraction! — Function

function addInteraction!(agentModel::Model, addvar::Symbol, addeqs::String)

Add a interaction parameters to the model. Differently to the local interactions, this parameters will be updated inside the integration steps of the differential equations.

Examples

m = Model();
eq = "
dxdt = -x+g*ξ #Wiener process with an additional interaction
"
addVariable!(m,:x,eq);

interaction = "
g₁ += 1./sqrt((x₁-x₂)^2+(y₁-y₂)^2) #The difussion will be affected by the presence of other particles around them.
"
addInteraction!(m,:g,interaction);

function addInteraction!(agentModel::Model, addvar::Symbol, addeqs::String)

Add a interaction parameters to the model. Differently to the local interactions, this parameters will be updated inside the integration steps of the differential equations.

Examples

m = Model();
eq = "
dxdt = -x+g*ξ+p #Wiener process with an additional interaction
"
addVariable!(m,:x,eq);

interaction = "
g₁ += 1./sqrt((x₁-x₂)^2+(y₁-y₂)^2) #The difussion will be affected by the presence of other particles around them.
p₁ += 1./(abs(x₁-x₂)+abs(y₁-y₂)) #The difussion will be affected by the presence of other particles around them.
"
addInteraction!(m,[:g,:p],interaction);

Neighborhoods

AgentModel.setNeighborhoodFull! — Function

function setNeighborhoodFull!(agentModel::Model)

Function that sets the neighborhoodof the model to fully connected. It is the neighborhood by default if no other neighborhood is defined.

AgentModel.setNeighborhoodAdjacency! — Function

function setNeighborhoodAdjacency!(agentModel::Model, condition::String, nMax::Int)

Function that creates a track of neighbours based depending on a condition. It also requires an estimation of the maximum amount of neighbours that a particle will have, the program will crash if during the evolution, the particle finds more than nMax particles fullfilling the condition.

Examples

m = Model()

addLocal!([:x,:y])

condition= #Euler distance
"
sqrt((x₁-x₂)^2+(y₁-y₂)^2) < 1.
"
setNeighborhoodAdjacency!(m,condition,nMax=50) #No more than 50 particles will be in the neighborhood at any time

AgentModel.setNeighborhoodGrid! — Function

function setNeighborhoodGrid!(agentModel::Model, vars::Array{Symbol}, box::Matrix{<:AbstractFloat}, radius::AbstractFloat)

Function that keeps track of the cells by assotiating them to a grid position. The neighborhood requires that the variables that define the dimensions of the grid are defined, the size of each dimension and the radius of interaction of each particle.

Example

m = Model()

addLocal!([:x,:y])

vars = [:x,:y] #The particles are neighbors depending on the x and y variables.
box = [[-1,-1],[1,1]] #The particles are in a square of size 2 around zero.
radius = 0.5 #The particles interact at most with particles 0.5 far appart from them.
setNeighborhoodGrid!(m,vars,box,radius)

function setNeighborhoodGrid!(agentModel::Model, vars::Array{Symbol}, box::Matrix{<:AbstractFloat}, radius::Array{<:AbstractFloat})

Function that keeps track of the cells by assotiating them to a grid position with diferent range for each dimension. The neighborhood requires that the variables that define the dimensions of the grid are defined, the size of each dimension and the radius of interaction of each particle.

Example

m = Model()

addLocal!([:x,:y])

vars = [:x,:y] #The particles are neighbors depending on the x and y variables.
box = [[-1,-1],[1,1]] #The particles are in a square of size 2 around zero.
radius = [0.5,1.] #The particles interact assymetrically in both dimensions.
setNeighborhoodGrid!(m,vars,box,radius)

Special

AgentModel.addDivision! — Function

function addDivision!(agentModel::Model, condition::String, update::String; randVar = Tuple{Symbol,String}[])

Function that adds a division process of the particles to the model. Cells divide under condition and update the new parameters with update.

Examples

m = Model();
addLocal!(m,[:g,:tDiv]); #Add two local variables, g and the time of division tDiv

condition = 
"
tDiv < t
"

update =
"
g₁ = gₚ*r #Asymmetric split of the component between the two daughter particles
g₁ = tDivₚ+1.

g₂ = gₚ*(1-r)
g₂ = tDivₚ+1.
"

addDivision!(m,condition,update,randVar=[(:r,"Uniform",0.,1.)])

AgentModel.addPseudopode! — Function

function addPseudopode!(agentModel::Model, var::Symbol, tVar::Symbol, neighbourCondition::String, force::String, updateChange::String; randVar = Tuple{Symbol,String}[])

Add a pseudopode force.

Examples

m = Model()
addLocal!([:x,:y])

neighbourCondition = 
"
sqrt((x₁-x₂)^2+(y₁-y₂)^2) < 2.
"
force = 
"
f = sqrt((x₁-x₂)^2+(y₁-y₂)^2)*exp(-sqrt((x₁-x₂)^2+(y₁-y₂)^2))
"
updateChange = 
"
tPseudo = t + σPseudo
"
addPseudopode!(m, :f, :tPseudo, condition, force, updateChange, randVar = [(:σPseudo,"Uniform",1.,2.)])