API for Developers

Add parameters

AgentModel.parameterAdapt — Function

function parameterAdapt(agentModel::Model,inLoop,arg;platform::String="cpu",nChange_=false)

Function that returns the pieces of the final compiling code for the parameters adapted to the corresponding platform:

arrays to declare containg the parameters declared

functions for parameter updates

execution lines

Parameters:

agentModel : Model structure

inLoop : Code of the interaction local to be adapted depending on the neighbborhood

arg : Additional arguments required by the functions

platform : Platform to be adapted ("cpu" or "gpu")

nChange_ : FILL THE GAP

Set neighbourhood

All the neighborhood types X share the same structure. They contain always four elements:

A NeighborhoodX structure that keeps all the required elements in order to define the neighborhood. This structured is created when calling the setNeighborhoodX! and added to the model.

The setNeighborhoodX! function that is the high level function for seting a neighborhood. The details are present in the API.

A neighboursX function that is runned during compilation time and creates the functions that computed to make the neighborhoods.

A neighboursXAdapt that assigns the inner loop iterator :nnic2_ into the appropiate term depending on the neighborhood type.

Functions defined in 2 and 4 have always the same inputs and outputs.

Full connected

AgentModel.NeighboursFull — Type

struct NeighboursFull <: Neighbours

AgentModel.neighboursFull — Function

function neighboursFull(agentModel::Model;platform="cpu")

AgentModel.neighboursFullAdapt — Function

function neighboursFullAdapt(entry)

By Adjacecy

AgentModel.NeighboursAdjacency — Type

struct NeighboursAdjacency <: Neighbours

AgentModel.neighboursByAdjacency — Function

function neighboursByAdjacency(agentModel::Model;platform="cpu")

AgentModel.neighboursByAdjacencyAdapt — Function

function neighboursByAdjacencyAdapt(entry)

By Grid

AgentModel.NeighboursGrid — Type

struct NeighboursGrid <: Neighbours

AgentModel.neighboursByGrid — Function

function neighboursByGrid(agentModel::Model;platform="cpu")

AgentModel.neighboursByGridAdapt — Function

function neighboursByGridAdapt(entry)

AgentModel.loopNeighbourGridCreation — Function

function loopNeighbourGridCreation(i,i0,n,x=nothing,pos="")

Auxiliar function for creating nested loops during the grid creation.

Auxiliar

AgentModel.subs — Function

function subs(expr,ob,tar)

Substitute ob expression or symbol by tar expression or symbol into expression expr.

function subs(expr,ob,tar)

Substitute ob expression or symbol by tar expression or symbol into expression expr.