API for Developers

Add parameters

Missing docstring.

Missing docstring for parameterAdapt. Check Documenter's build log for details.

Set neighbourhood

All the neighborhood types X share the same structure. They contain always four elements:

A NeighborhoodX structure that keeps all the required elements in order to define the neighborhood. This structured is created when calling the setNeighborhoodX! and added to the model.

The setNeighborhoodX! function that is the high level function for seting a neighborhood. The details are present in the API.

A neighboursX function that is runned during compilation time and creates the functions that computed to make the neighborhoods.

A neighboursXAdapt that assigns the inner loop iterator :nnic2_ into the appropiate term depending on the neighborhood type.

Functions defined in 2 and 4 have always the same inputs and outputs.

Full connected

Missing docstring.

Missing docstring for NeighboursFull. Check Documenter's build log for details.

Missing docstring.

Missing docstring for neighboursFull. Check Documenter's build log for details.

Missing docstring.

Missing docstring for neighboursFullAdapt. Check Documenter's build log for details.

By Adjacecy

Missing docstring.

Missing docstring for NeighboursAdjacency. Check Documenter's build log for details.

Missing docstring.

Missing docstring for neighboursByAdjacency. Check Documenter's build log for details.

Missing docstring.

Missing docstring for neighboursByAdjacencyAdapt. Check Documenter's build log for details.

By Grid

Missing docstring.

Missing docstring for NeighboursGrid. Check Documenter's build log for details.

Missing docstring.

Missing docstring for neighboursByGrid. Check Documenter's build log for details.

Missing docstring.

Missing docstring for neighboursByGridAdapt. Check Documenter's build log for details.

Missing docstring.

Missing docstring for loopNeighbourGridCreation. Check Documenter's build log for details.

Auxiliar

AgentBasedModels.subs — Function

function subs(expr,ob,tar)

Substitute ob expression or symbol by tar expression or symbol into expression expr.

function subs(expr,ob,tar)

Substitute ob expression or symbol by tar expression or symbol into expression expr.